requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.7.1
Django==5.1.2
django-allauth==0.61.1
django-environ==0.11.2
Faker==30.8.1
idna==3.10
oauthlib==3.2.2
pillow==10.4.0
psycopg==3.2.3
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
